+++ PayPalDataCollector/src/main/java/com/braintreepayments/api/PayPalDataCollectorInternalRequest.java
@@ -7,17 +7,22 @@ import androidx.annotation.NonNull;
 /**
  * Used to configuration the PayPalDataCollector request
  */
-class PayPalDataCollectorRequest {
+class PayPalDataCollectorInternalRequest {
 
     private String applicationGuid;
     private String clientMetadataId;
     private boolean disableBeacon;
+    final private boolean hasUserLocationConsent;
     private HashMap<String,String> additionalData;
+
+    PayPalDataCollectorInternalRequest(boolean hasUserLocationConsent) {
+        this.hasUserLocationConsent = hasUserLocationConsent;
+    }
 
     /**
      * @param additionalData Additional data that should be associated with the data collection.
      */
-    PayPalDataCollectorRequest setAdditionalData(HashMap<String, String> additionalData) {
+    PayPalDataCollectorInternalRequest setAdditionalData(HashMap<String, String> additionalData) {
         this.additionalData = additionalData;
 
         return this;
@@ -27,7 +32,7 @@ class PayPalDataCollectorRequest {
      * @param applicationGuid The application global unique identifier.
      *                        There is a 36 character length limit on this value.
      */
-    PayPalDataCollectorRequest setApplicationGuid(String applicationGuid) {
+    PayPalDataCollectorInternalRequest setApplicationGuid(String applicationGuid) {
         this.applicationGuid = applicationGuid;
 
         return this;
@@ -36,7 +41,7 @@ class PayPalDataCollectorRequest {
     /**
      * @param riskCorrelationId The desired pairing ID, trimmed to 32 characters.
      */
-    PayPalDataCollectorRequest setRiskCorrelationId(@NonNull String riskCorrelationId) {
+    PayPalDataCollectorInternalRequest setRiskCorrelationId(@NonNull String riskCorrelationId) {
         this.clientMetadataId = riskCorrelationId.substring(0, Math.min(riskCorrelationId.length(), 32));
 
         return this;
@@ -45,7 +50,7 @@ class PayPalDataCollectorRequest {
     /**
      * @param disableBeacon Indicates if the beacon feature should be disabled.
      */
-    PayPalDataCollectorRequest setDisableBeacon(boolean disableBeacon) {
+    PayPalDataCollectorInternalRequest setDisableBeacon(boolean disableBeacon) {
         this.disableBeacon = disableBeacon;
 
         return this;
@@ -61,6 +66,10 @@ class PayPalDataCollectorRequest {
 
     String getClientMetadataId() {
         return clientMetadataId;
+    }
+
+    public boolean getHasUserLocationConsent() {
+        return hasUserLocationConsent;
     }
 
     boolean isDisableBeacon() {